slackbot/pyproject.toml
[project]
name = "slackbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=22.0.0",
    "google-cloud-aiplatform>=1.138.0",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-vertexai>=3.2.2",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "slack-bolt>=1.27.0",
    "snowflake-connector-python>=3.12.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "peft>=0.18.1",
    "mcp[cli]>=1.26.0",
    "requests>=2.32.5",
    "psycopg[binary]>=3.1",
    "sqlparse>=0.5.0",
]
